requirements.txt
python-telegram-bot
diffusers
transformers
torch
accelerate
python-dotenv